Add hg64_put(), the counterpart of hg64_get()
To import data from elsewhere, in general we need to be given the
range of values covering a bucket in the source data, and a count of
the values occurring in that bucket. This range might cover more than
one counter in our histogram.

The hg64_put() function has been pulled out of hg64_merge() and
generalized to allow for buckets that do not line up. I have also
thought of a better way to cope with remainders.